Create separate file DateUtils containing class DateVisualTransformation and fun formatDateString and Implement changes in AddMeeting also in EditMeeting
