+++ app/src/main/java/com/swent/suddenbump/ui/calendar/AddMeeting.kt
@@ -13,8 +13,6 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.platform.testTag
-import androidx.compose.ui.text.AnnotatedString
-import androidx.compose.ui.text.TextRange
 import androidx.compose.ui.text.TextStyle
 import androidx.compose.ui.text.input.TextFieldValue
 import androidx.compose.ui.unit.dp
@@ -25,9 +23,8 @@ import com.swent.suddenbump.model.user.UserViewModel
 import com.swent.suddenbump.ui.navigation.NavigationActions
 import java.text.SimpleDateFormat
 import java.util.*
-import androidx.compose.ui.text.input.VisualTransformation
-import androidx.compose.ui.text.input.OffsetMapping
-import androidx.compose.ui.text.input.TransformedText
+import com.swent.suddenbump.ui.utils.DateVisualTransformation
+import com.swent.suddenbump.ui.utils.formatDateString
 
 
 @OptIn(ExperimentalMaterial3Api::class)
@@ -111,7 +108,9 @@ fun AddMeetingScreen(
                             dateFormat.isLenient = false
                             val parsedDate = dateFormat.parse(date.text)
                             val calendar = GregorianCalendar().apply {
-                                time = parsedDate
+                                if (parsedDate != null) {
+                                    time = parsedDate
+                                }
                                 set(Calendar.HOUR_OF_DAY, 0)
                                 set(Calendar.MINUTE, 0)
                                 set(Calendar.SECOND, 0)
@@ -124,7 +123,7 @@ fun AddMeetingScreen(
                                 friendId = friendId,
                                 location = location,
                                 date = meetingDate,
-                                creatorId = userViewModel.getCurrentUser().value?.uid ?: "",
+                                creatorId = userViewModel.getCurrentUser().value.uid,
                                 accepted = false
                             )
                             meetingViewModel.addMeeting(newMeeting)
@@ -171,64 +170,4 @@ fun AddMeetingScreen(
     )
 }
 
-fun formatDateString(input: TextFieldValue): TextFieldValue {
-    val digitsOnly = input.text.filter { it.isDigit() } // Extract only digits
-    val formatted = StringBuilder()
 
-    // Add slashes at correct positions: after day (2nd digit) and month (4th digit)
-    for (i in digitsOnly.indices) {
-        formatted.append(digitsOnly[i])
-        if ((i == 1 || i == 3) && i < digitsOnly.length - 1) {
-            formatted.append('/')
-        }
-    }
-
-    // Adjust the cursor position for the added slashes
-    val originalCursorPosition = input.selection.start
-    val slashCountBeforeCursor = input.text.take(originalCursorPosition).count { it == '/' }
-    val expectedSlashCount = formatted.count { it == '/' }
-
-    // Update cursor position based on added/removed slashes
-    var newCursorPosition = originalCursorPosition + (expectedSlashCount - slashCountBeforeCursor)
-    newCursorPosition = newCursorPosition.coerceIn(0, formatted.length)
-
-    return TextFieldValue(formatted.toString(), TextRange(newCursorPosition))
-}
-
-class DateVisualTransformation : VisualTransformation {
-    override fun filter(text: AnnotatedString): TransformedText {
-        val trimmed = text.text.filter { it.isDigit() } // Keep only digits
-        val formatted = StringBuilder()
-
-        // Build formatted text with slashes
-        for (i in trimmed.indices) {
-            formatted.append(trimmed[i])
-            if ((i == 1 || i == 3) && i < trimmed.length - 1) {
-                formatted.append('/')
-            }
-        }
-
-        // Map cursor positions
-        val offsetMapping = object : OffsetMapping {
-            override fun originalToTransformed(offset: Int): Int {
-                return when {
-                    offset <= 1 -> offset
-                    offset <= 3 -> offset + 1
-                    offset <= 6 -> offset + 2
-                    else -> offset + 2
-                }.coerceAtMost(formatted.length)
-            }
-
-            override fun transformedToOriginal(offset: Int): Int {
-                return when {
-                    offset <= 2 -> offset
-                    offset <= 5 -> offset - 1
-                    offset <= 8 -> offset - 2
-                    else -> offset - 2
-                }.coerceAtMost(trimmed.length)
-            }
-        }
-
-        return TransformedText(AnnotatedString(formatted.toString()), offsetMapping)
-    }
-}
